Wirego package cleanup
